// Nextjs/authentication/src/lib/dtos/auth.dto.ts
import { z } from 'zod';

export const registerSchema = z.object({
  username: z.string().min(3, 'Username must be at least 3 characters long').max(20, 'Username must be at most 20 characters long'),
  email: z.string().email('Invalid email address'),
  password: z.string().min(6, 'Password must be at least 6 characters long'),
});

export const loginSchema = z.object({
  email: z.string().email('Invalid email address'),
  password: z.string().min(1, 'Password is required'),
});

export type RegisterUserDTO = z.infer<typeof registerSchema>;
export type LoginUserDTO = z.infer<typeof loginSchema>;